Move classes to a common ui package to be reused by other verticals
